requirements.txt
langchain
faiss-cpu 
sentence-transformers
langchain-community
huggingface_hub
streamlit
python-dotenv